+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `StableVec::insert_into_hole()`
 - `StableVec::grow()`
 - `StableVec::clear()`
+- `StableVec::from_vec()`
+- `StableVec::extend_from_slice()`
+- `Debug` implementations for `Iter`, `IterMut` and `Keys`
+- `Write` implementation for `StableVec<u8>`
 
 ### Changed
 - The `Drop` impl now uses the `mem::needs_drop()` optimization hint to avoid